Parallel Loop doesn't work iterating through maps
