+++ cloudslang-runtime/src/main/java/io/cloudslang/lang/runtime/bindings/ParallelLoopBinding.java
@@ -10,20 +10,15 @@
 package io.cloudslang.lang.runtime.bindings;
 
 import io.cloudslang.lang.entities.LoopStatement;
-import io.cloudslang.lang.entities.MapForLoopStatement;
 import io.cloudslang.lang.entities.SystemProperty;
 import io.cloudslang.lang.entities.bindings.values.Value;
-import io.cloudslang.lang.entities.bindings.values.ValueFactory;
 import io.cloudslang.lang.runtime.bindings.scripts.ScriptEvaluator;
 import io.cloudslang.lang.runtime.env.Context;
-import java.io.Serializable;
-import java.util.ArrayList;
+
 import java.util.List;
-import java.util.Map;
 import java.util.Set;
 import org.apache.commons.collections4.CollectionUtils;
 import org.apache.commons.lang3.Validate;
-import org.apache.commons.lang3.tuple.Pair;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
@@ -54,20 +49,14 @@ public class ParallelLoopBinding extends AbstractBinding {
         Validate.notNull(systemProperties, "system properties cannot be null");
         Validate.notNull(nodeName, "node name cannot be null");
 
-        List<Value> result = new ArrayList<>();
+        List<Value> result;
         try {
             Value evalResult = scriptEvaluator.evalExpr(parallelLoopStatement.getExpression(),
                     flowContext.getImmutableViewOfVariables(), systemProperties, parallelLoopStatement.getFunctionDependencies());
 
             evalResult = getEvalResultForMap(evalResult, parallelLoopStatement, parallelLoopStatement.getExpression());
 
-            if (evalResult != null && evalResult.get() != null) {
-                //noinspection unchecked
-                for (Serializable serializable : ((List<Serializable>) evalResult.get())) {
-                    Value value = ValueFactory.create(serializable, evalResult.isSensitive());
-                    result.add(value);
-                }
-            }
+            result = (List<Value>) getIterableFromEvalResult(evalResult);
         } catch (Throwable t) {
             throw new RuntimeException(generateParallelLoopExpressionMessage(nodeName, t.getMessage()), t);
         }
